Se resuelve el error que tenía con las pruebas en la importaciones, se resuelve la tarea dejada por el profesor

+++ 04-multiplication/src/domain/use-cases/save-file.use-case.test.ts
@@ -1,12 +1,11 @@
-import { describe, expect, it, test } from '@jest/globals';
-// import '@types/jest'; si lo agrego el editor de codigo me marca rojo el afterEach
 import { SaveFile } from './save-file.use-case';
 import * as fs from 'fs';
 
 describe('SaveFileUseCase', () => {
 
-    afterEach(() => {
-        fs.rmSync('outputs', { recursive: true, force: true });
+    afterEach(() => {       
+        const outputExists = fs.existsSync('outputs');
+        if( outputExists ) fs.rmSync('outputs', { recursive: true, force: true });
     });
     
 
@@ -22,6 +21,26 @@ describe('SaveFileUseCase', () => {
         const checkFile = fs.existsSync(filePath);
         const fileContent = fs.readFileSync(filePath, { encoding: 'utf-8' });
         
+        expect( result ).toBeTruthy();
+        expect( checkFile ).toBeTruthy();
+        expect( fileContent ).toEqual(options.fileContent);
+    });
+
+    test('Should save file with custom values', () => {
+        
+        const saveFile = new SaveFile();
+        const options = {
+            fileContent: 'custom content',
+            fileDestination: 'outputs',
+            fileName: 'custom-table-name'
+        }
+
+        const filePath = `${options.fileDestination}/${options.fileName}.txt`;
+        
+        const result = saveFile.execute(options);
+        const checkFile = fs.existsSync(filePath);
+        const fileContent = fs.readFileSync(filePath, { encoding: 'utf-8' });
+
         expect( result ).toBeTruthy();
         expect( checkFile ).toBeTruthy();
         expect( fileContent ).toEqual(options.fileContent);
